Create fox_10.0 branch (work-in-progress); full credits to @bigbiff, @mauronofio, @Captain_Throwback, and all the official TWRP team for their Android-10.0 branch

+++ prebuilt/relink.sh
@@ -4,18 +4,16 @@ process_file()
 {
     dst=$1/$(basename $2)
     src=$2
+
+    [[ -e $src ]] || return 0
+
     if [ $dst == $src ]; then
       cp -f -p $src $src.tmp
       src=$2.tmp
-    else
-      cp -f -p $src $dst
     fi
 
-    sed "s|/system/bin/linker64\x0|/sbin/linker64\x0\x0\x0\x0\x0\x0\x0|g" $src | sed "s|/system/bin/linker\x0|/sbin/linker\x0\x0\x0\x0\x0\x0\x0|g" | sed "s|/system/bin/sh\x0|/sbin/sh\x0\x0\x0\x0\x0\x0\x0|g" | sed "s|/system/lib64\x0|/sbin\x0\x0\x0\x0\x0\x0\x0\x0\x0|g" | sed "s|/system/lib\x0|/sbin\x0\x0\x0\x0\x0\x0\x0|g" > $dst
-
-    if [ $1 == $(dirname $2) ]; then
-      rm -f $src
-    fi
+    cp $src $dst
+    if [[ -e $2.tmp ]]; then rm -f $2.tmp; fi
 }
